Refractor main file, add otion file
